fix(Voice): better Enum naming
